pyproject.toml
[tool.poetry]
name = "examtex"
version = "0.1.0"
description = "Generate exams in tex"
authors = ["Patrick Toale <ptoale@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"
jinja2 = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
